requirements.txt
absl-py==0.11.0
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.8
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
audioread==2.1.9
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Cython==0.29.22
dataclasses==0.6
decorator==5.0.5
fairseq==0.10.2
fsspec==0.8.7
future==0.18.2
google-auth==1.27.0
google-auth-oauthlib==0.4.2
grpcio==1.36.0
hydra-core==1.0.6
idna==2.10
importlib-metadata==3.7.0
importlib-resources==5.1.1
joblib==1.0.1
librosa==0.8.0
llvmlite==0.36.0
Markdown==3.3.4
multidict==5.1.0
numba==0.53.1
numpy==1.20.1
oauthlib==3.1.0
omegaconf==2.0.6
packaging==20.9
pandas==1.2.2
Pillow==8.1.1
pooch==1.3.0
portalocker==2.2.1
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==1.2.3
pytorch-ranger==0.1.1
pytz==2021.1
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
sacrebleu==1.5.0
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
SoundFile==0.10.3.post1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
torch==1.7.1
torch-optimizer==0.1.0
torchaudio==0.7.2
torchvision==0.8.2
tqdm==4.58.0
typing-extensions==3.7.4.3
urllib3==1.26.3
wavencoder==0.1.2
Werkzeug==1.0.1
yarl==1.6.3
zipp==3.4.0
